[Feat] 무기 강화 알고리즘
기본적인 알고리즘 구현

+++ Assets/02. Scripts/UserData/UserData.cs
@@ -54,9 +54,9 @@ public class Weapon
 {
     public int index { get; set; }
     public string name { get; set; }
-    public double price { get; set; }
+    public long price { get; set; }
     public float probability { get; set; }
-    public double enhancingPrice { get; set; }
+    public long enhancingPrice { get; set; }
     public string addressID { get; set; }
     public List<int> needItems { get; set; }
     private Sprite sprite;
